# EmuKey/backend/openapi/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/activations/challenge": {
      "post": {
        "operationId": "LicensingController_challenge",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ActivationChallengeDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DeviceChallengeDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Create a device challenge; activation challenges resolve the license from the bearer activation key and require no purchaser login",
        "tags": [
          "licensing"
        ]
      }
    },
    "/api/v1/activations": {
      "post": {
        "operationId": "LicensingController_activate",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ActivateDeviceDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LicenseDeviceDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Activate a device with a bearer activation key and device proof; no purchaser session is required",
        "tags": [
          "licensing"
        ]
      }
    },
    "/api/v1/licenses/action-verification": {
      "post": {
        "operationId": "LicensingController_requestActionVerification",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LicensingActionVerificationDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "licensing"
        ]
      }
    },
    "/api/v1/licenses/action-verification/resolve": {
      "post": {
        "operationId": "LicensingController_resolveActionVerification",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ResolveLicensingActionDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LicensingActionResolutionDto"
                }
              }
            }
          },
          "401": {
            "description": ""
          },
          "403": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Resolve the licensing action an email token authorizes without consuming it",
        "tags": [
          "licensing"
        ]
      }
    },
    "/api/v1/commands/{commandId}": {
      "get": {
        "operationId": "LicensingController_commandStatus",
        "parameters": [
          {
            "name": "commandId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Phase6CommandStatusDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "licensing"
        ]
      }
    },
    "/api/v1/licenses/{licenseId}/devices/{deviceId}/revoke": {
      "post": {
        "operationId": "LicensingController_revokeDevice",
        "parameters": [
          {
            "name": "licenseId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "deviceId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RevokeDeviceDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LicenseDeviceDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "licensing"
        ]
      }
    },
    "/api/v1/licenses/{licenseId}/devices/{deviceId}/remote-revoke": {
      "post": {
        "operationId": "LicensingController_remoteRevokeDevice",
        "parameters": [
          {
            "name": "licenseId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "deviceId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RemoteRevokeDeviceDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LicenseDeviceDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "licensing"
        ]
      }
    },
    "/api/v1/licenses/{licenseId}/activation-key/rotate": {
      "post": {
        "operationId": "LicensingController_rotate",
        "parameters": [
          {
            "name": "licenseId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RotateActivationKeyDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Phase6CommandDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "licensing"
        ]
      }
    },
    "/api/v1/licenses/{licenseId}/activation-key/recover": {
      "post": {
        "operationId": "LicensingController_recoverActivationKey",
        "parameters": [
          {
            "name": "licenseId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ActivationKeyRecoveryDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Phase6CommandDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "licensing"
        ]
      }
    },
    "/api/v1/licenses/{licenseId}/lifecycle": {
      "post": {
        "operationId": "LicensingController_lifecycle",
        "parameters": [
          {
            "name": "licenseId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LicenseLifecycleDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Phase6CommandDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "licensing"
        ]
      }
    },
    "/api/v1/entitlements/issue": {
      "post": {
        "operationId": "LicensingController_issueEntitlement",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EntitlementRefreshDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EntitlementDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "licensing"
        ]
      }
    },
    "/api/v1/entitlements/refresh": {
      "post": {
        "operationId": "LicensingController_refreshEntitlement",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EntitlementRefreshDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EntitlementDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "licensing"
        ]
      }
    },
    "/api/v1/entitlements/verify": {
      "post": {
        "operationId": "LicensingController_verifyEntitlement",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EntitlementVerifyDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EntitlementValidationDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "licensing"
        ]
      }
    }
  },
  "info": {
    "title": "Emukey API",
    "description": "Versioned API for the Emukey license platform",
    "version": "1.0.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "LicenseDeviceDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "deviceRef": {
            "type": "string"
          },
          "status": {
            "type": "string",
            "enum": [
              "ACTIVE",
              "REVOKED"
            ]
          },
          "bindingGeneration": {
            "type": "number"
          },
          "activatedAt": {
            "type": "object",
            "nullable": true
          },
          "revokedAt": {
            "type": "object",
            "nullable": true
          }
        },
        "required": [
          "id",
          "deviceRef",
          "status",
          "bindingGeneration"
        ]
      },
...
      "ActivationChallengeDto": {
        "type": "object",
        "properties": {
          "licenseId": {
            "type": "string",
            "format": "uuid",
            "description": "License id for device-bound entitlement or revoke challenges; not required for activation."
          },
          "activationKey": {
            "type": "string",
            "pattern": "^0x[0-9a-fA-F]{64}$",
            "description": "Required for activation challenges; used to resolve the license without customer ownership."
          },
          "purpose": {
            "type": "string",
            "enum": [
              "ACTIVATE_DEVICE",
              "SELF_REVOKE_DEVICE",
              "ISSUE_ENTITLEMENT",
              "REFRESH_ENTITLEMENT"
            ]
          },
          "deviceRef": {
            "type": "string",
            "minLength": 1,
            "maxLength": 128
          },
          "deviceId": {
            "type": "string",
            "format": "uuid",
            "description": "Set when requesting a challenge to revoke an existing device"
          }
        },
        "required": [
          "purpose",
          "deviceRef"
        ]
      },
      "DeviceChallengeDto": {
        "type": "object",
        "properties": {
          "licenseId": {
            "type": "string",
            "format": "uuid",
            "description": "Resolved license id; returned only after the activation credential has been accepted."
          },
          "challenge": {
            "type": "string"
          },
          "expiresAt": {
            "type": "string",
            "format": "date-time"
          },
          "bindingGeneration": {
            "type": "number"
          },
          "keyVersion": {
            "type": "number"
          },
          "purpose": {
            "type": "string",
            "enum": [
              "ACTIVATE_DEVICE",
              "SELF_REVOKE_DEVICE",
              "ISSUE_ENTITLEMENT",
              "REFRESH_ENTITLEMENT"
            ]
          }
        },
        "required": [
          "licenseId",
          "challenge",
          "expiresAt",
          "bindingGeneration",
          "keyVersion",
          "purpose"
        ]
      },
      "ActivateDeviceDto": {
        "type": "object",
        "properties": {
          "activationKey": {
            "type": "string",
            "pattern": "^0x[0-9a-fA-F]{64}$",
            "description": "Bearer activation credential supplied by the enterprise administrator"
          },
          "challenge": {
            "type": "string"
          },
          "deviceRef": {
            "type": "string",
            "minLength": 1,
            "maxLength": 128
          },
          "devicePublicKey": {
            "type": "string",
            "description": "EVM address corresponding to the device signing key"
          },
          "proof": {
            "type": "string",
            "description": "EIP-191 signature over the activation challenge"
          }
        },
        "required": [
          "activationKey",
          "challenge",
          "deviceRef",
          "devicePublicKey",
          "proof"
        ]
      },
      "LicensingActionVerificationDto": {
        "type": "object",
        "properties": {
          "action": {
            "type": "string",
            "enum": [
              "ROTATE_KEY",
              "REVOKE_DEVICE",
              "REMOTE_REVOKE_DEVICE",
              "KEY_RECOVERY"
            ]
          },
          "licenseId": {
            "type": "string",
            "format": "uuid"
          },
          "deviceId": {
            "type": "string",
            "format": "uuid",
            "description": "Required for device-scoped revoke actions"
          }
        },
        "required": [
          "action",
          "licenseId"
        ]
      },
      "ResolveLicensingActionDto": {
        "type": "object",
        "properties": {
          "actionToken": {
            "type": "string",
            "minLength": 20
          }
        },
        "required": [
          "actionToken"
        ]
      },
      "LicensingActionResolutionDto": {
        "type": "object",
        "properties": {
          "action": {
            "type": "string",
            "enum": [
              "ROTATE_KEY",
              "REVOKE_DEVICE",
              "REMOTE_REVOKE_DEVICE",
              "KEY_RECOVERY"
            ]
          },
          "licenseId": {
            "type": "string",
            "format": "uuid"
          },
          "deviceId": {
            "type": "string",
            "format": "uuid",
            "nullable": true
          },
          "expiresAt": {
            "type": "string",
            "format": "date-time"
          }
        },
        "required": [
          "action",
          "licenseId",
          "expiresAt"
        ]
      },
      "Phase6CommandStatusDto": {
        "type": "object",
        "properties": {
          "commandId": {
            "type": "string",
            "format": "uuid"
          },
          "status": {
            "type": "string",
            "enum": [
              "PENDING",
              "SUBMITTED",
              "SUBMITTED_UNKNOWN",
              "CONFIRMED",
              "RETRYABLE_FAILED",
              "DEAD_LETTER",
              "ABANDONED",
              "SUPERSEDED"
            ]
          },
          "licenseId": {
            "type": "string",
            "format": "uuid"
          },
          "deviceId": {
            "type": "string",
            "format": "uuid",
            "nullable": true
          },
          "commandType": {
            "type": "string",
            "enum": [
              "ISSUE_LICENSE",
              "RENEW_LICENSE",
              "SUSPEND_LICENSE",
              "RESUME_LICENSE",
              "REVOKE_LICENSE",
              "ROTATE_KEY",
              "SYNC_DEVICE_COUNT"
            ]
          },
          "confirmedAt": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "transactionHash": {
            "type": "string",
            "pattern": "^0x[0-9a-fA-F]{64}$",
            "nullable": true
          }
        },
        "required": [
          "commandId",
          "status",
          "licenseId",
          "commandType"
        ]
      },
      "RevokeDeviceDto": {
        "type": "object",
        "properties": {
          "actionToken": {
            "type": "string",
            "minLength": 32
          },
          "activationKey": {
            "type": "string",
            "pattern": "^0x[0-9a-fA-F]{64}$"
          },
          "challenge": {
            "type": "string"
          },
          "proof": {
            "type": "string",
            "description": "EIP-191 signature over the revoke challenge"
          }
        },
        "required": [
          "actionToken",
          "activationKey",
          "challenge",
          "proof"
        ]
      },
      "RemoteRevokeDeviceDto": {
        "type": "object",
        "properties": {
          "actionToken": {
            "type": "string",
            "minLength": 32
          },
          "currentPassword": {
            "type": "string",
            "minLength": 1
          }
        },
        "required": [
          "actionToken",
          "currentPassword"
        ]
      },
      "RotateActivationKeyDto": {
        "type": "object",
        "properties": {
          "actionToken": {
            "type": "string",
            "minLength": 32
          },
          "currentKey": {
            "type": "string",
            "pattern": "^0x[0-9a-fA-F]{64}$"
          }
        },
        "required": [
          "actionToken",
          "currentKey"
        ]
      },
      "Phase6CommandDto": {
        "type": "object",
        "properties": {
          "commandId": {
            "type": "string",
            "format": "uuid"
          },
          "status": {
            "type": "string",
            "enum": [
              "PENDING",
              "SUBMITTED",
              "SUBMITTED_UNKNOWN",
              "CONFIRMED",
              "RETRYABLE_FAILED",
              "DEAD_LETTER",
              "ABANDONED",
              "SUPERSEDED"
            ]
          },
          "licenseId": {
            "type": "string",
            "format": "uuid"
          },
          "deviceId": {
            "type": "string",
            "format": "uuid",
            "nullable": true
          }
        },
        "required": [
          "commandId",
          "status",
          "licenseId"
        ]
      },
      "ActivationKeyRecoveryDto": {
        "type": "object",
        "properties": {
          "actionToken": {
            "type": "string",
            "minLength": 32
          },
          "currentPassword": {
            "type": "string",
            "minLength": 1
          }
        },
        "required": [
          "actionToken",
          "currentPassword"
        ]
      },
      "LicenseLifecycleDto": {
        "type": "object",
        "properties": {
          "command": {
            "type": "string",
            "enum": [
              "SUSPEND_LICENSE",
              "RESUME_LICENSE",
              "REVOKE_LICENSE"
            ]
          },
          "reason": {
            "type": "string"
          }
        },
        "required": [
          "command"
        ]
      },
      "EntitlementRefreshDto": {
        "type": "object",
        "properties": {
          "licenseId": {
            "type": "string",
            "format": "uuid"
          },
          "deviceId": {
            "type": "string",
            "format": "uuid"
          },
          "challenge": {
            "type": "string",
            "minLength": 16
          },
          "proof": {
            "type": "string",
            "description": "EIP-191 device signature over the entitlement challenge"
          }
        },
        "required": [
          "licenseId",
          "deviceId",
          "challenge",
          "proof"
        ]
      },
      "EntitlementDto": {
        "type": "object",
        "properties": {
          "token": {
            "type": "string"
          },
          "expiresAt": {
            "type": "string",
            "format": "date-time"
          },
          "licenseId": {
            "type": "string",
            "format": "uuid"
          },
          "deviceId": {
            "type": "string",
            "format": "uuid"
          },
          "entitlementVersion": {
            "type": "number"
          }
        },
        "required": [
          "token",
          "expiresAt",
          "licenseId",
          "deviceId",
          "entitlementVersion"
        ]
      },
      "EntitlementVerifyDto": {
        "type": "object",
        "properties": {
          "token": {
            "type": "string",
            "description": "Signed entitlement JWT returned by issue or refresh"
          }
        },
        "required": [
          "token"
        ]
      },
      "EntitlementValidationDto": {
        "type": "object",
        "properties": {
          "bindingGeneration": {
            "type": "number"
          },
          "valid": {
            "type": "boolean",
            "enum": [
              true
            ]
          },
          "expiresAt": {
            "type": "string",
            "format": "date-time"
          },
          "licenseId": {
            "type": "string",
            "format": "uuid"
          },
          "deviceId": {
            "type": "string",
            "format": "uuid"
          },
          "entitlementVersion": {
            "type": "number"
          },
          "keyVersion": {
            "type": "number"
          },
          "rights": {
            "type": "object",
            "additionalProperties": true
          }
        },
        "required": [
          "bindingGeneration",
          "valid",
          "expiresAt",
          "licenseId",
          "deviceId",
          "entitlementVersion",
          "keyVersion",
          "rights"
        ]
      }
    }
  }
}